# src/main/java/com/example/solvaProject/model/dto/AccountDto.java
package com.example.solvaProject.model.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class AccountDto {
    private String accountNumber;
    private double balanceUsd;
    private double limitUsd;
}